requirements.txt
fastapi==0.115.5
langchain-core==0.3.21
langchain-ollama==0.2.1
langchain-openai==0.2.10
pydantic==2.10.2
Requests==2.32.3
streamlit==1.40.2
python-dotenv==1.0.1